Refactor ContractsTable and introduce formatting utilities
- Enhanced ContractsTable by importing new formatting utilities for ETH amounts, file sizes, and dates, improving data presentation.
- Removed redundant local formatting functions and utilized centralized formatting functions from the new utils module.
- Implemented a dynamic loading color effect during data fetching to enhance user experience.
- Adjusted pagination limit in the useContracts hook to optimize data handling.
